Working Eclipse Plugin :confetti_ball:
Migrated proof of concept fix to a stand-alone plugin.  I'm able to run
this off of the 3.0.1-dev build of the standard findbugs eclipse plugin.

I'm opting to go for a condensed package configuration because I don't
want to have to deal with package names like
edu.umd.cs.findbugs.plugin.eclipse.quickfix.exception, for example.