:confetti_ball: Able to use fb-contrib's code
This commit and bec6e70ce80ffdb4f7f932563ab107edc3752443 enable both the
built version of the plugin and the "launch configuration" to work.

In this commit I, (with some parallel changes in
kjlubick/fb-contrib@94f5ca73bf52df485339454270fc09e03b5c15d2) allowed this Eclipse plugin to
optionally require the OSGI bundle "fb-contrib-6.1.0" and then updated
the launch configuration to include them.

This allows the launch configuration to work, meaning we can debug/run
without having to build and install/update in a test environment.

The requirement is optional because we won't install a fb-contrib
Eclipse plugin (other than this one, that is) - we use the compiled in
resources available in the last commit.

This setup is (now fully) mirrored after the FindBugs Eclipse plugin
build scheme.